refactor: 스토리북 상호작용 테스트 코드 분리
- 각 컴포넌트 디렉터리의 [컴포넌트].play.js 파일로 분리
