+++ src/components/Molecules/ExpiredDateInput/ExpiredDataInput.stories.jsx
@@ -1,11 +1,8 @@
 import React from 'react';
-import { expect } from '@storybook/jest';
-import { within, userEvent } from '@storybook/testing-library';
 import ExpiredDateInput from '.';
 import useExpiredDate from 'hooks/useExpiredDate';
-import { PLACEHOLDER } from 'constant';
 import MESSAGE from 'constant/message';
-import { CLASS } from 'constant/selector';
+import { validInputCasePlay, invalidDateInputCasePlay } from './ExpiredDataInput.play';
 
 export default {
   title: 'Molecules/ExpiredDateInput',
@@ -54,44 +51,8 @@ export const Default = Template.bind({});
 
 export const ValidInputCase = Template.bind({});
 
-ValidInputCase.play = async ({ canvasElement }) => {
-  // given
-  const canvas = within(canvasElement);
-  const monthInput = canvas.getByPlaceholderText(PLACEHOLDER.MONTH);
-  const yearInput = canvas.getByPlaceholderText(PLACEHOLDER.YEAR);
-
-  const INPUTTED_MONTH = '12';
-  const INPUTTED_YEAR = '22';
-
-  const EXPECTED_BLANK_MESSAGE = '';
-
-  // when
-  await userEvent.type(monthInput, INPUTTED_MONTH);
-  await userEvent.type(yearInput, INPUTTED_YEAR);
-
-  // then
-  const invalidMessage = document.getElementsByClassName(CLASS.INVALID_INPUT_MESSAGE)[0];
-
-  await expect(invalidMessage.textContent).toBe(EXPECTED_BLANK_MESSAGE);
-};
+ValidInputCase.play = validInputCasePlay;
 
 export const InvalidDateInputCase = Template.bind({});
 
-InvalidDateInputCase.play = async ({ canvasElement }) => {
-  // given
-  const canvas = within(canvasElement);
-  const monthInput = canvas.getByPlaceholderText(PLACEHOLDER.MONTH);
-  const yearInput = canvas.getByPlaceholderText(PLACEHOLDER.YEAR);
-
-  const INPUTTED_MONTH = '12';
-  const INPUTTED_YEAR = '20';
-
-  // when
-  await userEvent.type(monthInput, INPUTTED_MONTH);
-  await userEvent.type(yearInput, INPUTTED_YEAR);
-
-  // then
-  const invalidMessage = document.getElementsByClassName(CLASS.INVALID_INPUT_MESSAGE)[0];
-
-  await expect(invalidMessage.textContent).toBe(MESSAGE.INVALID_EXPIRED_DATE);
-};
+InvalidDateInputCase.play = invalidDateInputCasePlay;
