+++ src/components/Molecules/PasswordInput/PasswordInput.stories.jsx
@@ -1,10 +1,8 @@
 import React from 'react';
-import { expect } from '@storybook/jest';
-import { userEvent } from '@storybook/testing-library';
 import PasswordInput from '.';
 import usePassword from 'hooks/usePassword';
 import MESSAGE from 'constant/message';
-import { CLASS } from 'constant/selector';
+import { validInputCasePlay, emptyPasswordInputPlay } from './PasswordInput.play';
 
 export default {
   title: 'Molecules/PasswordInput',
@@ -56,34 +54,8 @@ export const Default = Template.bind({});
 
 export const ValidInputCase = Template.bind({});
 
-ValidInputCase.play = async () => {
-  // given
-  const passwordInputs = Array.from(document.getElementsByClassName(CLASS.INPUT_PASSWORD));
-  const INPUTTED_PASSWORD = ['1', '2'];
-
-  // when
-  passwordInputs.forEach(async (passwordInput, index) => {
-    await userEvent.type(passwordInput, INPUTTED_PASSWORD[index]);
-  });
-
-  // then
-  const invalidMessage = document.getElementsByClassName(CLASS.INVALID_INPUT_MESSAGE)[0];
-
-  await expect(invalidMessage.textContent).toBe('');
-};
+ValidInputCase.play = validInputCasePlay;
 
 export const EmptyPasswordInput = Template.bind({});
 
-EmptyPasswordInput.play = async () => {
-  // given
-  const passwordInput = document.getElementsByClassName(CLASS.INPUT_PASSWORD)[1];
-  const INPUTTED_PASSWORD = '2';
-
-  // when
-  await userEvent.type(passwordInput, INPUTTED_PASSWORD);
-
-  // then
-  const invalidMessage = document.getElementsByClassName(CLASS.INVALID_INPUT_MESSAGE)[0];
-
-  await expect(invalidMessage.textContent).toBe(MESSAGE.INVALID_PASSWORD);
-};
+EmptyPasswordInput.play = emptyPasswordInputPlay;
