+++ src/components/Molecules/SecurityNumberInput/SecurityNumberInput.stories.jsx
@@ -1,10 +1,8 @@
 import React from 'react';
-import { expect } from '@storybook/jest';
-import { userEvent } from '@storybook/testing-library';
 import SecurityNumberInput from '.';
 import useSecurityNumber from 'hooks/useSecurityNumber';
 import MESSAGE from 'constant/message';
-import { CLASS } from 'constant/selector';
+import { validInputCasePlay, notTypeThreeNumberPlay } from './SecurityNumberInput.play';
 
 export default {
   title: 'Molecules/SecurityNumberInput',
@@ -56,32 +54,8 @@ export const Default = Template.bind({});
 
 export const ValidInputCase = Template.bind({});
 
-ValidInputCase.play = async () => {
-  // given
-  const securityNumberInput = document.getElementsByClassName(CLASS.INPUT_SECURITY_NUMBER)[0];
-  const INPUTTED_SECURITY_NUMBER = '123';
-
-  // when
-  await userEvent.type(securityNumberInput, INPUTTED_SECURITY_NUMBER);
-
-  // then
-  const invalidMessage = document.getElementsByClassName(CLASS.INVALID_INPUT_MESSAGE)[0];
-
-  await expect(invalidMessage.textContent).toBe('');
-};
+ValidInputCase.play = validInputCasePlay;
 
 export const NotTypeThreeNumber = Template.bind({});
 
-NotTypeThreeNumber.play = async () => {
-  // given
-  const securityNumberInput = document.getElementsByClassName(CLASS.INPUT_SECURITY_NUMBER)[0];
-  const INPUTTED_SECURITY_NUMBER = '12';
-
-  // when
-  await userEvent.type(securityNumberInput, INPUTTED_SECURITY_NUMBER);
-
-  // then
-  const invalidMessage = document.getElementsByClassName(CLASS.INVALID_INPUT_MESSAGE)[0];
-
-  await expect(invalidMessage.textContent).toBe(MESSAGE.INVALID_SECURITY_NUMBER);
-};
+NotTypeThreeNumber.play = notTypeThreeNumberPlay;
